requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
asgiref==3.7.2
    # via django
crispy-bootstrap5==2024.2
    # via -r .\requirements.in
django==5.0.3
    # via
    #   -r .\requirements.in
    #   crispy-bootstrap5
    #   django-crispy-forms
    #   django-extensions
    #   djangorestframework
django-crispy-forms==2.1
    # via
    #   -r .\requirements.in
    #   crispy-bootstrap5
django-environ==0.11.2
    # via -r .\requirements.in
django-extensions==3.2.3
    # via -r .\requirements.in
djangorestframework==3.14.0
    # via -r .\requirements.in
pytz==2024.1
    # via djangorestframework
sqlparse==0.4.4
    # via django
tzdata==2024.1
    # via django
